+++ tomcat/src/main/java/com/techcourse/controller/LoginController.java
@@ -9,8 +9,8 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Optional;
-import org.apache.catalina.ResourceResolver;
 import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.request.ResourceResolver;
 import org.apache.coyote.http.cookie.HttpCookie;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
