+++ tomcat/src/main/java/com/techcourse/controller/StaticFileController.java
@@ -5,8 +5,8 @@ import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import org.apache.catalina.ResourceResolver;
 import org.apache.catalina.controller.AbstractController;
+import org.apache.catalina.request.ResourceResolver;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.response.HttpResponse;
