+++ tomcat/src/main/java/org/apache/catalina/request/ResourceResolver.java
@@ -1,4 +1,4 @@
-package org.apache.catalina;
+package org.apache.catalina.request;
 
 import java.net.URL;
