+++ tomcat/src/main/java/org/apache/coyote/http/handler/RequestHandler.java
@@ -6,8 +6,8 @@ import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Paths;
-import org.apache.catalina.ResourceResolver;
 import org.apache.catalina.controller.Controller;
+import org.apache.catalina.request.ResourceResolver;
 import org.apache.coyote.http.request.HttpRequest;
 import org.apache.coyote.http.request.RequestLine;
 import org.apache.coyote.http.response.HttpResponse;
